Fix missing av disabled threads, move av_open to central place, add flv support
